# Replace built-in MPH with a dependency on the 'phf' crate