--- a/Cargo.toml
+++ b/Cargo.toml
@@ -34,6 +34,9 @@ default = []
 unstable = []
 no_std = []
 
+[dependencies]
+phf = { version = "0.11", default-features = false }
+
 [dev-dependencies.unicode_names2_macros]
 path = "unicode_names2_macros"
 version = ">=0.3, <0.7"