Before:

```
$ cargo bench
test tests::character_10000    ... bench:      95,820 ns/iter (+/- 3,313)
test tests::character_basic    ... bench:         188 ns/iter (+/- 4)
test tests::name_10000_invalid ... bench:      35,322 ns/iter (+/- 237)
test tests::name_all_valid     ... bench:   7,546,306 ns/iter (+/- 24,243)
test tests::name_basic         ... bench:          78 ns/iter (+/- 12)
$ ll target/release/deps/unicode_names2-a53ae0f01b2261d9
-rwxr-xr-x 1 dev-rust dev-rust 21M  6 août  12:21 target/release/deps/unicode_names2-a53ae0f01b2261d9

$ RUSTFLAGS="-C target-cpu=native" cargo bench
test tests::character_10000    ... bench:      88,172 ns/iter (+/- 2,905)
test tests::character_basic    ... bench:         189 ns/iter (+/- 4)
test tests::name_10000_invalid ... bench:      33,011 ns/iter (+/- 589)
test tests::name_all_valid     ... bench:   7,561,671 ns/iter (+/- 47,556)
test tests::name_basic         ... bench:          89 ns/iter (+/- 6)
$ ll target/release/deps/unicode_names2-a53ae0f01b2261d9
-rwxr-xr-x 1 dev-rust dev-rust 21M  6 août  12:24 target/release/deps/unicode_names2-a53ae0f01b2261d9
```

After:

```
$ cargo bench
test tests::character_10000    ... bench:     100,275 ns/iter (+/- 1,521)
test tests::character_basic    ... bench:         162 ns/iter (+/- 7)
test tests::name_10000_invalid ... bench:      35,341 ns/iter (+/- 258)
test tests::name_all_valid     ... bench:   7,480,702 ns/iter (+/- 79,935)
test tests::name_basic         ... bench:          84 ns/iter (+/- 13)
$ ll target/release/deps/unicode_names2-a7967b9cdf0f0db1
-rwxr-xr-x 1 dev-rust dev-rust 26M  6 août  12:21 target/release/deps/unicode_names2-a7967b9cdf0f0db1

$ RUSTFLAGS="-C target-cpu=native" cargo bench
test tests::character_10000    ... bench:      97,400 ns/iter (+/- 3,124)
test tests::character_basic    ... bench:         169 ns/iter (+/- 4)
test tests::name_10000_invalid ... bench:      34,721 ns/iter (+/- 5,037)
test tests::name_all_valid     ... bench:   7,590,151 ns/iter (+/- 83,504)
test tests::name_basic         ... bench:          92 ns/iter (+/- 10)
$ ll target/release/deps/unicode_names2-a7967b9cdf0f0db1
-rwxr-xr-x 1 dev-rust dev-rust 26M  6 août  12:23 target/release/deps/unicode_names2-a7967b9cdf0f0db1
```

while running on i5-1240P (downclocked to 2GHz)

so performance is similar, but `character_10000` is noticeably slower, and the binary size grew by 5MB, which is not negligeable.

@youknowone Could you try rewriting #12 based on this? I'd like to know the performance and size difference between your current code (keeps the custom MPH, and adds a `phf` MPH just for the aliases) and using a single `phf` MPH for both tables.